+++ src/WebhookerServiceProvider.php
@@ -8,10 +8,14 @@ use Illuminate\Support\ServiceProvider;
 use TechraysLabs\Webhooker\Commands\CircuitResetCommand;
 use TechraysLabs\Webhooker\Commands\CircuitStatusCommand;
 use TechraysLabs\Webhooker\Commands\CleanExpiredSecretsCommand;
+use TechraysLabs\Webhooker\Commands\DeadLetterCommand;
 use TechraysLabs\Webhooker\Commands\EndpointDisableCommand;
 use TechraysLabs\Webhooker\Commands\EndpointEnableCommand;
 use TechraysLabs\Webhooker\Commands\EndpointListCommand;
 use TechraysLabs\Webhooker\Commands\HealthCommand;
+use TechraysLabs\Webhooker\Commands\HealthSnapshotCommand;
+use TechraysLabs\Webhooker\Commands\PartitionCreateCommand;
+use TechraysLabs\Webhooker\Commands\PartitionDropCommand;
 use TechraysLabs\Webhooker\Commands\PruneCommand;
 use TechraysLabs\Webhooker\Commands\ReplayCommand;
 use TechraysLabs\Webhooker\Commands\SecretRotateCommand;
@@ -21,14 +25,17 @@ use TechraysLabs\Webhooker\Contracts\InboundProcessor;
 use TechraysLabs\Webhooker\Contracts\PayloadValidator;
 use TechraysLabs\Webhooker\Contracts\RetryStrategy;
 use TechraysLabs\Webhooker\Contracts\SignatureGenerator;
+use TechraysLabs\Webhooker\Contracts\WebhookLock;
 use TechraysLabs\Webhooker\Contracts\WebhookMetrics;
 use TechraysLabs\Webhooker\Contracts\WebhookRepository;
 use TechraysLabs\Webhooker\Services\CacheCircuitBreaker;
+use TechraysLabs\Webhooker\Services\CacheLockProvider;
 use TechraysLabs\Webhooker\Services\ConfigPayloadValidator;
 use TechraysLabs\Webhooker\Services\DefaultInboundProcessor;
 use TechraysLabs\Webhooker\Services\EloquentWebhookMetrics;
-use TechraysLabs\Webhooker\Services\EloquentWebhookRepository;
 use TechraysLabs\Webhooker\Services\HmacSignatureGenerator;
+use TechraysLabs\Webhooker\Storage\PartitionManager;
+use TechraysLabs\Webhooker\Storage\WebhookStorageManager;
 use TechraysLabs\Webhooker\Strategies\ExponentialBackoffRetry;
 
 class WebhookerServiceProvider extends ServiceProvider
@@ -40,7 +47,15 @@ class WebhookerServiceProvider extends ServiceProvider
     {
         $this->mergeConfigFrom(__DIR__.'/../config/webhooks.php', 'webhooks');
 
-        $this->app->bind(WebhookRepository::class, EloquentWebhookRepository::class);
+        // Storage driver manager (resolves WebhookRepository via driver pattern)
+        $this->app->singleton(WebhookStorageManager::class, function ($app) {
+            return new WebhookStorageManager($app);
+        });
+
+        $this->app->bind(WebhookRepository::class, function ($app) {
+            return $app->make(WebhookStorageManager::class)->driver();
+        });
+
         $this->app->bind(RetryStrategy::class, function ($app) {
             $config = $app['config']->get('webhooks.retry', []);
 
@@ -55,6 +70,12 @@ class WebhookerServiceProvider extends ServiceProvider
         $this->app->bind(WebhookMetrics::class, EloquentWebhookMetrics::class);
         $this->app->bind(CircuitBreaker::class, CacheCircuitBreaker::class);
         $this->app->bind(PayloadValidator::class, ConfigPayloadValidator::class);
+
+        // Distributed locking for horizontal scaling
+        $this->app->bind(WebhookLock::class, CacheLockProvider::class);
+
+        // Partition manager
+        $this->app->singleton(PartitionManager::class);
     }
 
     /**
@@ -85,6 +106,10 @@ class WebhookerServiceProvider extends ServiceProvider
                 __DIR__.'/../resources/views' => resource_path('views/vendor/webhooker'),
             ], 'webhooker-views');
 
+            $this->publishes([
+                __DIR__.'/../database/stubs' => database_path('stubs/webhooker'),
+            ], 'webhooker-stubs');
+
             $this->commands([
                 PruneCommand::class,
                 ReplayCommand::class,
@@ -97,6 +122,10 @@ class WebhookerServiceProvider extends ServiceProvider
                 SimulateCommand::class,
                 SecretRotateCommand::class,
                 CleanExpiredSecretsCommand::class,
+                DeadLetterCommand::class,
+                HealthSnapshotCommand::class,
+                PartitionCreateCommand::class,
+                PartitionDropCommand::class,
             ]);
         }
     }